+++ examples/common/universal+redux/components/app.js
@@ -1,22 +1,21 @@
 import {Layout} from '../../components/layout';
-import React, {Component, PropTypes} from 'react';
+import React, {PropTypes} from 'react';
 import {connect} from 'react-redux';
 
-class App extends Component {
-  render () {
-    const {children, server} = this.props;
+function AppComponent (props) {
+  const {children, server} = props;
 
-    return <Layout isRedux server={server}>{children}</Layout>;
-  }
+  return <Layout isRedux server={server}>{children}</Layout>;
 }
 
-App.propTypes = {
+AppComponent.propTypes = {
   children: PropTypes.node,
+  location: PropTypes.shape({}),
   server: PropTypes.string
 };
 
 function mapStateToProps (state) {
-  return {server: state.server};
+  return {location: state.location, server: state.server};
 }
 
-export default connect(mapStateToProps)(App);
+export const App = connect(mapStateToProps)(AppComponent);